+++ example/lib/main.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'dart:core';
-import 'src/basic_sample/basic_sample.dart';
-import 'src/call_sample/call_sample.dart';
+import 'src/loopback_sample.dart';
+import 'src/get_user_media_sample.dart';
+import 'src/data_channel_sample.dart';
 import 'src/route_item.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 void main() => runApp(new MyApp());
 
@@ -19,13 +19,10 @@ enum DialogDemoAction {
 
 class _MyAppState extends State<MyApp> {
   List<RouteItem> items;
-  String _serverAddress = '192.168.31.152';
-  SharedPreferences prefs;
 
   @override
   initState() {
     super.initState();
-    _initData();
     _initItems();
   }
 
@@ -57,78 +54,33 @@ class _MyAppState extends State<MyApp> {
     );
   }
 
-  _initData() async {
-    prefs = await SharedPreferences.getInstance();
-    setState(() {
-      _serverAddress = prefs.getString('server') ?? '';
-    });
-  }
-
-  void showDemoDialog<T>({BuildContext context, Widget child}) {
-    showDialog<T>(
-      context: context,
-      builder: (BuildContext context) => child,
-    ).then<void>((T value) {
-      // The value passed to Navigator.pop() or null.
-      if (value != null) {
-        if (value == DialogDemoAction.connect) {
-          prefs.setString('server', _serverAddress);
-          Navigator.push(
-              context,
-              MaterialPageRoute(
-                  builder: (BuildContext context) =>
-                      CallSample(ip: _serverAddress)));
-        }
-      }
-    });
-  }
-
-  _showAddressDialog(context) {
-    showDemoDialog<DialogDemoAction>(
-        context: context,
-        child: new AlertDialog(
-            title: const Text('Enter server address:'),
-            content: TextField(
-              onChanged: (String text) {
-                setState(() {
-                  _serverAddress = text;
-                });
-              },
-              decoration: InputDecoration(
-                hintText: _serverAddress,
-              ),
-              textAlign: TextAlign.center,
-            ),
-            actions: <Widget>[
-              new FlatButton(
-                  child: const Text('CANCEL'),
-                  onPressed: () {
-                    Navigator.pop(context, DialogDemoAction.cancel);
-                  }),
-              new FlatButton(
-                  child: const Text('CONNECT'),
-                  onPressed: () {
-                    Navigator.pop(context, DialogDemoAction.connect);
-                  })
-            ]));
-  }
-
   _initItems() {
     items = <RouteItem>[
       RouteItem(
-          title: 'Basic API Tests',
-          subtitle: 'Basic API Tests.',
+          title: 'GetUserMedia',
           push: (BuildContext context) {
             Navigator.push(
                 context,
                 new MaterialPageRoute(
-                    builder: (BuildContext context) => new BasicSample()));
+                    builder: (BuildContext context) =>
+                        new GetUserMediaSample()));
           }),
       RouteItem(
-          title: 'P2P Call Sample',
-          subtitle: 'P2P Call Sample.',
+          title: 'LoopBack Sample',
           push: (BuildContext context) {
-            _showAddressDialog(context);
+            Navigator.push(
+                context,
+                new MaterialPageRoute(
+                    builder: (BuildContext context) => new LoopBackSample()));
+          }),
+      RouteItem(
+          title: 'DataChannel',
+          push: (BuildContext context) {
+            Navigator.push(
+                context,
+                new MaterialPageRoute(
+                    builder: (BuildContext context) =>
+                        new DataChannelSample()));
           }),
     ];
   }