+++ CHANGELOG.md
@@ -1,3 +1,18 @@
-## [0.0.1] - 2018.05.30
+## Changelog
+--------------------------------------------
+[0.0.3] - 2018.12.20
+
+* Update WebRTC to 1.0.25821.
+* Implemented MediaStreamTrack.setVolume().
+* Add public getter for texture id.
+* Fixed getUserMedia does not work for capturing audio only.
+
+[0.0.2] - 2018.11.04
+
+* Add 'enabled' method for MediaStreamTrack.
+* Implement switch camera.
+* Add arm64-v8a and x86_64 architecture support for android.
+
+[0.0.1] - 2018.05.30
 
 * Initial release.